The screen works!
The screen drawing functions work as well as the background color. Also I added a small compile script for local compiling, when developing code (compile.sh). In the top folder there is also a small script for clean, compile and flash the game files to the board (game_flash.sh).